Started removing ProcessResponses

+++ MyPortal/Processes/SystemProcesses.cs
@@ -183,7 +183,12 @@ namespace MyPortal.Processes
 
         public static bool ValidateUpn(string upn)
         {
-            var alpha = new [] { 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'J', 'K', 'L', 'M', 'N', 'P', 'Q', 'R', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z' };
+            var alpha = new[]
+            {
+                'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'J', 'K', 'L', 'M', 'N', 'P', 'Q', 'R', 'T', 'U', 'V', 'W', 'X',
+                'Y', 'Z'
+            };
+
             var chars = upn.ToCharArray();
 
             if (chars.Length != 13)
@@ -195,9 +200,9 @@ namespace MyPortal.Processes
 
             for (var i = 1; i < chars.Length; i++)
             {
-                var n = (int) char.GetNumericValue(chars[i]) * (i+1);
+                var n = char.GetNumericValue(chars[i]) * (i+1);
 
-                check += n;
+                check += (int) n;
             }
 
             var alphaIndex = check % 23;
